Remove old task packet code
